🚑 avoid rewritting user groups each time
The group was using (6, _, ids) which causes the groups to be removed
then added again even if they do not change. Groups usually do not
change, and some fields are recomputed after a group change (for example
the share field), so it is better to only add and remove groups that
really need to change.